requirements.txt
aiohttp>=3.3.0,<3.6.0
websockets>=6.0,<7.0
discord.py==1.2.4